requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
transparent-background==1.2.4
requests==2.31.0
